Added error handling for Sign Up, Sign In and Change Pass word forms.
Added logic to track when a user completes a form incorrectly and display errors on the front end. App now indicates if required fields are null, or if passwords do not match when creating an account or changing password.

Next steps: Track a user's session so change password and sign out can functionality can be completed.